requirements.txt
torch
torchvision
rasterio
numpy
Pillow
scikit-image
scikit-learn
piq
h5py
